#1536 move Config.cpp to config directory
