+++ core/src/config/handler/GpuSearchConfigHandler.cpp
@@ -9,12 +9,12 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 #ifdef MILVUS_GPU_VERSION
-#include "config/handler/GpuSearchConfigHandler.h"
 
 #include <string>
 #include <vector>
 
-#include "server/Config.h"
+#include "config/Config.h"
+#include "config/handler/GpuSearchConfigHandler.h"
 
 namespace milvus {
 namespace server {
