+++ core/src/wrapper/ConfAdapter.cpp
@@ -8,18 +8,18 @@
 // Unless required by applicable law or agreed to in writing, software distributed under the License
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
+
 #include "wrapper/ConfAdapter.h"
 
 #include <fiu-local.h>
-
 #include <cmath>
 #include <memory>
 #include <string>
 #include <vector>
 
 #include "WrapperException.h"
+#include "config/Config.h"
 #include "knowhere/index/vector_index/helpers/IndexParameter.h"
-#include "server/Config.h"
 #include "utils/Log.h"
 
 namespace milvus {
