+++ core/unittest/wrapper/test_knowhere.cpp
@@ -9,7 +9,7 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "server/Config.h"
+#include "config/Config.h"
 #include "wrapper/KnowhereResource.h"
 #include "wrapper/utils.h"
